+++ changelog.md
@@ -6,6 +6,7 @@
 > This version is not released yet and is under active development.
 
 - Fix highlighting of deprecated messages.
+- Use ASCII characters instead of unicode for prompt rendering in messages.
 
 ## [5.0.0 (2025-05-13)](https://github.com/kdeldycke/click-extra/compare/v4.15.0...v5.0.0)